replace skywire-deployment with skywire
